[bazel/otbn] Assemble source files into a single .o file
By passing multiple .s files to `riscv32-unknown-elf-as`, instead of
producing one .o file per .s, it is possible to make use of
conditional compilation using the `.set` directive. This is similar to
the preprocessor macros in C.

The Bazel build targets can be rewritten to having multiple source
inputs instead of declaring dependencies, e.g.,

```
otbn_sim_test(
    name = "rsa_512_dec_test",
    srcs = [
        "rsa_512_dec_test.s",
        "//sw/otbn/crypto:modexp.s",
        "//sw/otbn/crypto:montmul.s",
        "//sw/otbn/crypto:run_rsa_modexp_mem,s",
    ],
    testcase = "rsa_512_dec_test.hjson",
)
```

The main application for this feature is the conditional
enabling/disabling of hardening countermeasures. With the above test
this could look like this:

```
// rsa_512_dec_test.s
.set RSA_MODEXP_ENABLE_MESSAGE_BLINDING, 1

// modexp.s
.ifdef RSA_MODEXP_ENABLE_MESSAGE_BLINDING
  // Code with blinding
.else
  // Code without blinding
.endif
```

Signed-off-by: Andrea Caforio <andrea.caforio@lowrisc.org>