[sw,cov] Add on-device OTTF coverage runtime
This change adds a new on-device coverage runtime with OTTF transport.

When coverage is enabled, the runtime can collect profile counter data from the device and send it over the console.
On the host, these raw counter values can be combined with static profile data to recreate the full LLVM profile.

The coverage runtime consists of the following components:
* `api.h`: Provides a set of shared C APIs implemented by transport runtimes.
* `api_asm.h`: Provides APIs for assembly code.
* `printer.h`/`printer.c`: Implements the shared logic for different runtimes.
* Linker scripts (`bss.ld`, `info.ld`, `rodata.ld`): Define special sections for LLVM profile data, counters, and build ID.

OTTF transport implementation is also included:
* `ottf_runtime.c`: Integrates the coverage printer with the OTTF console for data output.
* `ottf_runtime_skip.c`: A no-op version of the runtime for builds where coverage is enabled but not instrumented.

Change-Id: I2e4b135a58b93a257fc37880700e23f4b99defc4
Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>