[sw,cryptolib] Follow OTBN style guide rules 1-4 related to shares
This commit changes all the instructions in the p384 library that
violate rules 1-4 related to the handling of shares.

Shares of the same secret should not be:
- in different parts of the same reg
- accessed in consecutive instructions
- accessed in the same instruction
- over write each other

In some parts of the p384 library the shares are combined and some
of these rules are violated by necessity.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>
(cherry picked from commit 89e99ebad133d4bdeaf7db2902cc55ccfbd36d53)