[dv] Remove unnecessary empty driver run_phase implementations
The fault here is in the template file that contained a trivial
implementation (that just called the superclass) and almost never
needed overriding.

There are a few drivers where this is changing the implementation
slightly:

  - uart_driver
  - usb20_driver

In both cases, I think the author believed that "reset signals" was
supposed to be called to reset the signals when a reset (or start of
simulation) happened. This isn't correct with the current repository,
but the rejig in this commit won't change behaviour at all.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>