[dv,alert] Get rid of two under_reset flags
This was all rather confusing! There were the following signals:

  alert_esc_agent_cfg::under_reset
  alert_esc_base_monitor::under_reset
  dv_base_driver::under_reset

And the alert and escalation side of things maintained them
differently(!): on the escalation side, under_reset meant "under
reset". On the alert side, it meant "has not been initialised since
reset", which is the same as !alert_esc_agent_cfg::alert_init_done.

This commit gets rid of all three flags, relying on
dv_base_agent_cfg::in_reset instead (which gets maintained properly by
the monitors).

Note that it's a little odd to ignore dv_base_driver::under_reset. It
won't matter with this commit (because dv_base_driver doesn't use that
flag directly and nothing subclasses the alert/escalation drivers),
but things are still a little strange. A following commit will get rid
of that flag entirely, but the alert agent was tricky, so I've done it
separately.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>