[bazel] blackhole stdin when running QEMU
Connecting `stdin` prevents QEMU from exiting (or using ^C to quit) when
run with `bazel run`. We don't normally want to talk to the OpenTitan
UART over stdin anyway, so I'm going to disable that.

When we have OpenTitanTool support we can communicate with the UART
through pipes separately from stdio.

Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>