[usbdev] Introduce packet transmission test mode
To assist device manufacturers with USB electrical compliance
testing, a test mode is introduced which causes the USB device
to transmit an IN DATA packet from Endpoint Zero repeatedly.

This test mode mimics the 'Test Packet' mode required of
High Speed USB devices, and may be used to transmit the same
sequence of bytes but with Full Speed signaling.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>