[dv,usbdev] Sequence for packet transmission test mode
Exercise packet transmission test mode. This is not a requirement
of the USB protocol specification for Full Speed devices. Rather
it is introduced to simplify electrical compliance testing.

The sequence therefore checks that packet transmission is occurring
and that the line is suitably busy, but does not concern itself
overly with the transmitted data being correct.

Monitor and scoreboard are not used in this sequence because the
traffic is not specification-compliant.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>