[bazel,qemu] Mark broken FPGA tests as broken in QEMU
These tests are broken in the fpga_cw310_sival_rom_ext exec env because
they are broken in the rom_ext envs in general, so they will also be
broken in QEMU. Mark these tests as broken for now to work around this.

Note - this commit is not exhaustive. It specifically addresses the
relevant `keymgr` tests as these were failures I looked into when
developing QEMU's keymgr.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>