[dv] Cleanup and document provided simulator .tcl in hw/dv/tools/
This commit overhauls the provided simulator .tcl scripts used to drive
supported RTL simulators at run-time. All .tcl code is moved to a new directory
hw/dv/tools/tcl, and is refactored into new libraries and heavily documented.
A new README.md file provides an overview of the use of this tooling, such as
what features exactly the default .tcl run-script provides, and how dvsim can
be used to force the use of a user-provided .tcl run script if desired. The
re-use of the provided .tcl libraries in custom run scripts is also explained.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>