[crypto] CTRL regs hardening: AES, GCM, HMAC, CSRNG, KeyMngr
Reread the CTRL and CTRL_AUX registers from the AES once it is done with its operation.
Move HMAC's verification of the CFG registers to the end of the computation.
Reread the CTRL register from the CSRNG after generating data.
Reread the CONTROL register from the key manager when generating a key.
This is to stop potential fault attacks on the security critical registers.

Signed-off-by: Siemen Dhooghe <sdhooghe@google.com>