[bazel] Disable OTT config file in modid check
Without this flag, OTT will pick up your local ~/.config/opentitan flags
and if those include an `--interface` it will try to connect and fail if
that's not found.

This causes problems if you have, for example, `--interface cw310` in
your config and don't have one connected when using the script.

Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>