[rescue,spidfu,test] Add SPI DFU state transition test
This commit introduces a new test case for SPI DFU rescue mode to verify
the correctness of DFU state machine transitions.

A new ROM_EXT configuration, `rom_ext_spidfu_flash_limit_zero`, is
intended to trigger a corner case where a write is attempted past the
end of the flash.

Signed-off-by: Anthony Chen <antchen@google.com>