[bazel,qemu] Support rom_ext binaries in the flashgen rule
Add support for loading two binaries (rom_ext and the boot firmware) as
that is the limit of what can be supported by QEMU's `flashgen.py`
currently, but can be used to support many `sival_rom_ext` tests. Add
explanatory comments and notes of limitations of `flashgen.py` where
relevant.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>