[hw/prim] Addition of prim_sdc_example for constraints checking
The module prim_sdc_example has been added to the prim repository. The
functionality of this block is irrelevant. It is only needed to run a
basic synthesis on a simplified design such that it can be verified that
all important synthesis constraints to preserve instances and prevent
logic optimization across preserved instances are applied correctly.

Signed-off-by: Michael Gautschi <mgautschi@lowrisc.org>