[rescue,test] Enhance Rescue GetBootLog test
This enhances the rescue GetBootLog test to improve flexibility in
specifying the ROM_EXT image and the active rom_ext boot slot.

This is useful in testing scenarios where the ROM_EXT is pre-flashed or
loaded outside of the `--bootstrap` command, allowing the user to
provide the expected binary for manifest comparison and specify the
boot slot.

Signed-off-by: Anthony Chen <antchen@google.com>
(cherry picked from commit 6802524d1de430d3cecde576ef455e9129fbd519)