[sw,cov] Set LSB in `coverage_status` to indicate validity
The existing check is indistinguishable from valid coverage reports if
`BUILD_ID` is also 0x42.

To fix this, this change always sets the LSB (Least Significant Bit) in
`coverage_status` when the coverage report is valid, while
`coverage_invalidate` is updated to set `coverage_status` to any
number without LSB to ensure it's invalid.

Change-Id: Ic8cd5d176d46333f1d9fec498df65c4301ca8f8b
Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>