[bazel,qemu] Get ROM from parameters instead of firmware
Corrects some of the QEMU execution environment Bazel logic so that it
fetches the ROM from the test parameters rather than the firmware,
allowing `qemu_params` to directly override the ROM, and supporting test
cases where `firmware` doesn't exist (e.g. when the firmware is the
result of some `opentitantool image assemble` command due to an
`assemble` param being used).

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>