[bazel,qemu] Add support for assembled test images
Allow QEMU tests to define an "assemble" param to create a custom image
to be loaded into flash for running the test. In the future this will
ideally be bootstrapped to match FPGA/silicon flows, but this is just
spliced into the flash image whilst boostrapping is not available.

Modifies the `gen_flash` FlashGen rule to instead take an assembled
binary (or just the provided firmware) and always store it at a location
of 0. This drops the limited support for using ELFs for convenient debug
symbols, but these should ideally not be used anyway due to the
mutability of flash meaning that the contents cannot be relied on (also,
loaded ELF symbols cannot be revoked for correctness in QEMU).

Also update to the latest QEMU release to support this, which contains
the relevant changes to QEMU's `flashgen.py`.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>