[bazel,qemu] Add QEMU envs for E2E dice chain tests
This commit adds QEMU environments for running the ROM_EXT end-to-end
dice chain tests.

Some of these test targets require bootstrapping, and as such need an
appropriate bootstrapping delay (e.g. 1 second) which is higher than the
default for QEMU environments, as a result of QEMU's slow emulation of
the ROM compared to FPGA/silicon. This is due to the SEC_MMIO hot loop
in the ROM falling in an unaligned (to a 4K page) EPMP region, which
causes QEMU'S TCG'S TLB caching to be unused.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>