[bazel,qemu] Add QEMU envs for E2E boot service tests
This commit only includes the non-ownership (activate, unlock) tests, as
the ownership tests are separated due to their need for more complicated
test infrastructure.

All tests are supported and passing in QEMU with the exception of the
`boot_service_wakeup` test, which is not expected to work, as QEMU's
pwrmgr currently does not model low power entry/exit.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>