[silicon_creator] Split chip_info data from headers
Having all in one is problematic because it can cause
a non-ROM binary to include the chip info symbol which
will then be placed somewhere else and override the
actual ROM chip info. The ROM should be the only one
including the chip info data whereas all other binaries
only need the header.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>