[otlib] Make Rescue 1.0 entry time configurable per-test
Also update the different ownership tests (ownership transfer, rescue
limit / permissions, new version and flash permission tests) to provide
an option for configuring this time, as it will need to be increased for
ownership tests to pass in QEMU environments.

This is because (even on a fast host) QEMU generally emulates the ROM
at a slower rate than FPGA/silicon targets due to a `sec_mmio` code hot
loop in the ROM falling in an EPMP region that is not aligned to a 4K
page, which means that QEMU's TCG cannot use its TLB, resulting in a
large performance decrease.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>