[bazel,qemu] Add QEMU envs for ROM_EXT E2E Ownership tests
Adds QEMU execution environment support for the ROM_EXT end-to-end
ownership tests, which test a variety of different ownership transfer
cases via the boot services.

To avoid large amounts of duplicated test definitions between the QEMU
and FPGA tests, extend the existing ownership test implementation to
instead take a dictionary of `shared_params` via which the FPGA and QEMU
params are then constructed. Dictionaries of additional parameters can
then be provided in the regular `fpga` and `qemu` fields which will
override the values in the shared parameters (allowing for per-env
customisation, or the addition of extra parameters).

A special case is added for handling the `test_cmd` to instead
automatically prepend an environment-specific prefix to the test
command to perform environment-specific setup (clear bitstream on FPGA,
increase rescue delay on QEMU). This is only applied if the `qemu/fpga`
parameter does not itself override the test command.

After several relevant fixes to QEMU and its Bazel integration, all 41
of these tests are currently passing in QEMU execution environments
(given a sufficiently fast host and not running in parallel, due to
the tight test timeouts).

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>