[manuf,perso] Don't manually control SpiConsole gpio pins for final hash
This is unnecessary, as the existing macros already control the sideband pins
and this confuses the HOST side since the DEVICE ends up indicating a read is
ready before any data has been put into the spi_device buffer. This can be
problematic in DV where the testbench can begin a write operation to fetch this
new data with the absolute minimum latency.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>