[dv,pwrmgr] fixing nightly failures
1. This fixes 'Offending '(esc_timeout || (!rst_lc_ni))''
signature in pwrmgr nightly runs.

2. clk_esc_i and clk_i are async to each other so a proper
fix is to model heart-beat of clk_esc_i and a counter that
keeps track of clk_esc_i being dead for 138 cycles of clk_i
and use that counter to trigger the assertion instead of
just using bare clk_esc_i value sampled on clk_i.

Signed-off-by: Phani Peri <pperi@lowrisc.org>