[sw,test] create QEMU I2C target mode test and harness using transport
This creates a test for Opentitan's I2C target mode for QEMU, using a
test harness that utilises the I2C QEMU transport implemented in the
previous commit.

The tested device emulates a simple "memory" device on the I2C bus.

The device consists of a 256-byte "memory". On a write transfer to the
device, the first byte sets the "pointer" to within the memory where
subsequent bytes are written to or read from. This value is incremented
after each written or read byte. During a read transfer, bytes from
memory to be read are buffered in the TX FIFO, which is refilled if its
fill level drops below a threshold.

The test harness uses the I2C transport device in QEMU to issue
transfers to write to and read from the emulated memory device.

Signed-off-by: Alice Ziuziakowska <a.ziuziakowska@lowrisc.org>