[sw,cryptolib] Hardened bn.sel for internal point mult
This commit adds hardening for the internal scalar point
multiplication.
Before the shares would be combined in the main loop to decide
whether 0, P or 2P should be added.
This commit determines this without combining the shares.
If the MSB of both shares are both 1 we want to add 2P.
If only one bit is 1 we want to add P.
If both bits are 0 we want to add 0.

In a simplified way this is done as explained below.

In a first step we pick P if MSB(d0) ^ MSB(d1) is 1 and 2P otherwise.
This is done as follows:
p_temp0 = d0 ? (d1 ? 2P : P) : (d1 ? P : 2P)

In a second step we pick between p_temp and 0.
If MSB(d0) | MSB(d1) = 1 then we pick p_temp otherwise we pick 0.
This is done as follows:
p_temp1 = d0 ? p_temp0 : 0
p_temp2 = d1 ? p_temp0 : p_temp1

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>
(cherry picked from commit 45453eced5e372e95f08aabbdc54a98bc43c0a83)