[sw,opentitanlib] QEMU I2C transport implementation
This implements a I2C transport for QEMU.

The transport uses a PTY to communicate with a `ot-i2c_host_proxy`
device to issue transfers to devices on the internal QEMU I2C buses.

See `docs/opentitan/i2c_host_proxy.md` in QEMU for more imformation
on the protocol.

Signed-off-by: Alice Ziuziakowska <a.ziuziakowska@lowrisc.org>