[rom_ext] Deref symlinks in imm_section release tarball
The `gen_stamped_tarball.py` script is used to create a release tarball
for the imm_section rom_ext. By default, `tarfile.open()` does not
dereference symbolic links, which means the tarball would contain broken
symlinks if the imm_section rom_ext are symlinks.

This change adds `dereference=True` to `tarfile.open()` to ensure that
symbolic links are followed and their target files are added to the
tarball instead of the symlinks themselves.

Change-Id: I17b2456b36e8efd411276b0702d2e430563e2d58
Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>