[rules,qemu] Make the qemu test runner accepts qemu arguments
This commit changes the qemu test script in two ways. First, it
is moved to a dedicated file. This has several advantages: it is
more readable and can be checked by shellcheck. Since it's a
template, I had to resort to some slight ugliness to make it work
but the overall result is at least quite readable.

Second, the test script parses its arguments are forwards then to
either qemu or the test harness. Specifically, any argument of the
form:
--qemu-arg=X
will forward X to qemu, and any argument of the form
--qemu-args=X,Y,Z
will forward X, Y, Z (and so on) as *different* arguments to qemu.

This allows one to do, for example:
bazel test //path/to/qemu:test --test_arg=--qemu-args=--trace,ot_usbdev*

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>