[opentitanlib] make `proxy_ops` return reference
It is a common pattern for dyn-safe objects to have additional capabilities to be
under an extension trait and have a method on the base trait to do the conversion.
However, in such cases, the extension trait should be implemented on the base type
itself and the dyn function should just be a `Ok(self)` return.

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>