[dv,ottf_spi_console] Extend ottf_spi_console.sv with async read-polling loop
Remove the task host_spi_console_read_wait_for() in favour of a new two-pronged
approach to capturing reads.
1) host_spi_console_poll_reads() is forked into a background process, which
automatically reads the spi_console when the DEVICE indicates new data is
available.
2) Sequences can call host_spi_console_read_wait_for_polled_str() to block
awaiting a certain string to be polled by the background task, at which point
the read buffer is flushed.

This allows for a much more tolerant testbench which will capture and ignore
messages from the DEVICE it is not expecting, while still allowing for a
stimulus vseq to block awaiting a particular message for synchronization
purposes.

Note.
A control bit 'enable_read_polling' is added to allow this background polling to
be paused temporarily while the existing task host_spi_console_read_payload()
takes over the task of issuing reads. This allows fixed payloads to be captured
using the synchronization messages and known test structrue as delimiters,
instead of needing to implement a message parser. See the code comment above
this bit for more details.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>