[sw,cov] Enable UartConsole matching mode for coverage builds
The `UartConsole` has a `uses_regex` method which controls whether
`UartConsole` enters a character-by-character read loop or not. The
intent for this character-by-character read loop is to avoid consuming
more output than necessary when a regular expression matches.

However, the `uses_regex` method only checks for `exit_success` and
`exit_failure` regexes. If the coverage mode is enabled, the
`UartConsole` also needs to read character-by-character to avoid
consuming output past coverage markers.

This commit renames `uses_regex` to `matching_mode` and makes it also
check for `ot_coverage_enabled` feature. This change ensures that
`UartConsole` will correctly enter a character-by-character read loop
when coverage is enabled.

Change-Id: I6544cf7e9a7b9fdf26fb8ff72fa684cf71867738
Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>