[opentitanlib] Add initial support for QEMU USB sense enable pin
This commit wires the necessary logic in the qemu transport to create
fake pin which controls the VBUS gate over the usb command chardev.
The fake pin uses number 254 and is aliases as "VBUS_SENSE_EN" which
is the named used by some FPGA tests for a similar VBUS gate.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>