[bazel,qemu] Handle non-assembled tests that specify custom firmware
Currently, if a test target provides custom device binaries (i.e. it
does not provide test source files), and does not specify an "assemble"
parameter for generating an image, the QEMU Bazel test dispatch rules do
not know how to handle this case. This is needed for example in the
`variation_interop_X_first_test` DICE Chain ROM_EXT end-to-end tests.

Add logic to correctly ignore assembly and changes to the "firmware"
param so that a blank image will be generated, and whatever binaries the
test defines for bootstrapping can then be used on the empty flash.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>