[rescue] Update rescue GPIO param to fix spidfu rescue test
The spidfu_rescue_boot_svc_req_disability test was introduced in
 #28385 and was passing its checks. However, a concurrent change
in #28440 renamed the RESCUE_GPIO parameter to WITH_RESCUE_GPIO_PARAM.

This renaming was not propagated to the test owner config for the
spidfu_rescue_boot_svc_req_disability, causing CI to keep failing on
spidfu_rescue_boot_svc_req_disability test and blocking the check.

This updates the parameter name to WITH_RESCUE_GPIO_PARAM in the
configuration to resolve the CI failures.

Signed-off-by: Anthony Chen <antchen@google.com>