[crypto] Add time constant memeq that works on bytes
The current hardened_memeq() is hardened against SCA and is const.
time. However, it requires that the input buffers are 32-bit word
aligned.

This commit adds a new consttime_memeq_byte() function that can
operate also on buffers that are not 32-bit aligned. It is const.
time but not hardened against SCA.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>