[cryptotest] Test RSA encrypt using wycheproof vectors
This commit extends the cryptotest framework to test the RSA encrypt
function. To test whether the encrypt was successful, the test
takes the ciphertext, decrypts it and compares it to the plaintext
that was fed into the encryption.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>
(cherry picked from commit 87279530bff70633c4e1be007afcd8e2480b81ba)