[opentitanlib] Introduce abstract over the USB backend
Currently there is a UsbBackend struct which wraps around rusb
and provides some utility functions. In the future, we may want
some transports to implement USB in a different way (e.g. QEMU
implements USB over a virtual channel). To do, the first step
is to create an abstraction over the existing `rusb` implementation.

This PR split UsbBackend into two: a context trait and a device trait.
The `rusb` implementation becomes just a particular implementation of
those traits.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>