[crypto] Use hardened_memcpy() in drbg
As mentioned in the issue, we should use hardened_memcpy() when copying
the seed. However, we only can use the hardened memcpy when the buffer
is 32-bit aligned. If this condition is not given, fall back to the
unhardened memcpy function.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>
(cherry picked from commit 55749867bdb8a55efcd26d50995910cdd1fc3541)