[rules,qemu] Always stop QEMU after the test
Even though opentitanlib supports asking QEMU to shutdown, it is
(1) optional, (2) relies on the test harness stopping cleanly.
One may argue that it is not really the responsability of the test
harness to shutdown QEMU anyway since it is not spawn by the harness
but rather by the qemu_test.sh script.

This commit makes the qemu_test.sh script always shutdown QEMU on
exit. It first sends a SIGINT and after one second a SIGKILL.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>