[fisim/crypto] Add instruction skip simulator
This only works for a debug enabled FPGA/chip and is only been tested on a CW340!

Add a script to load the pentest framework, connect to OpenOCD, and connect to GDB.
Generate a trace file of a called function.
Use the trace file to insert instruction skips in order to simulate fault attacks and test countermeasures.

Signed-off-by: Siemen Dhooghe <sdhooghe@google.com>