[docs,qemu] Add introductory QEMU documentation
Add some introductory documentation to explain QEMU, it's integration
into OpenTitan itself, and caveats that users should be aware of.

The focus of these docs are primarily on introducing the basic concepts
of QEMU, explaining the integration with OpenTitan, and noting any
caveats that are useful to be aware of when testing or developing
software stacks. Where possible, detailed explanations of QEMU's
internals or explanations of how OpenTitan emulation works is delegated
to the existing QEMU / QEMU fork documentation, to avoid replication and
docs getting out of date.

The one exception is the list of unemulated features - a short
non-comprehensive list is reproduced as it is probably useful that
anyone who wants to use QEMU is aware of some of the major caveats, even
if the list is not complete / comprehensive.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>