[flash_ctrl,dv] byte/word scrambling addr fix
FlashAddrWidth is in flash words (not bytes). Passing a byte address to
a function which expects word-addresses leads to a truncation of the three
most significant address bits. As a result, when accessing the upper part
of each flash bank the scrambling tweak was wrong which lead to a wrong
data encryption and hence also to a wrong ICV calculation for the expected
responses.

Signed-off-by: Michael Gautschi <mgautschi@lowrisc.org>
(cherry picked from commit d3aa4e9764fb730a0a1f61cf7b98ac3856914f22)