[utils] Add python script to find potential issues in netlist
This script parses the generated netlist for suspicious constructs and
counts the number of size_only instances. The script has been used with
one synthesis tool only and might need adjustments for other synthesis
tools. Further, the script relies on the preserved cells in the
technology specific prim library to be called "u_size_only_*".

Signed-off-by: Michael Gautschi <mgautschi@lowrisc.org>