[pentest] Add crypto version
Add the version of the crypto library to the pentest framework's readout
for the crypto related init functions.

In order to have the latest git hash be printed as the version, all
commands or tests should be run with the --stamp flag in the Bazel
command.

This is an automatic cherry-pick of #28620 to branch earlgrey_1.0.0.

Signed-off-by: Siemen Dhooghe <sdhooghe@google.com>