[adc_ctrl,dv] Put adc_ctrl_filter_cfg inside a proper class
This is inspired by trying to solve randomisation errors. It turns out
that randomising values that aren't in a *static* structure causes
some exciting problems with "solve before" constraints.

In hindsight, I think this change isn't quite needed: I could have
bodged something into the existing code. But I think it's probably an
improvement anyway (and I've got it working!).

One change is to make some of the class variable names a bit clearer.
For example, the "cond" field in adc_chn*_filter_ctl_* controls
whether the filter matches inside the range or outside of it. I
actually guessed wrong at first, but I think this makes a strong
argument for making the code more explicit on the DV side. That is now
reflected with a variable called "match_outside" (instead of my
original guess of "match_inside", which is exactly backwards!)

Indeed, I've just checked again and looked at
`theory_op_operation.md`. That document *does* mention a field called
"cond" and that it controls whether the filter matches inside or
outside of the range. It *doesn't* say which... This clearly needs
tidying up, but I think that sort of work can be folded into the next
proper release.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>