[adc_ctrl,dv] Fix path in call to DV_ASSERT_CTRL
The DV_ASSERT_CTRL macro expands into an $assertoff(...) call with the
hierarchy argument to choose where the assertions get turned off (or
on).

$assertoff() and $asserton() don't work with this style of path in all
cases. The LRM only requires them to work with local or absolute
paths: it doesn't allow up-references.

VCS *does* allow them (which is why the code was written like this),
but Xcelium doesn't. There are nice ways to do this cleanly, but they
are rather more effort than just shoving the absolute path of the dut
into the file here. Something to clean up when we come back to this
module, I think!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>