[cryptolib,silicon_creator] Deduplicate ibex random function
This commit makes two changes:
- the ibex_rnd32_read() function is moved from cryptolib to
  silicon_creator, the users are fixed
- the rnd_uint32() function is changed to return either mcycle
  OR rnd32(), depending on the boot stage and OTP setting.

Regarding the second change, it was previoulsy returning
`mcycle + rnd32()` but not waiting for random data depending on
an OTP setting. There is no good reason for that because if it waits
for data from the EDN, all the health checks are already done by the
csrng, meaning that adding `mcycle` does not add any value to a
random variable. Also if the OTP is *not* set, it causes a read
to known-invalid register using the fact that it will be 0 in this
cases.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>