[silicon_creator,ibex] Move code away from header
The code conditionally compiles from inline version of the code on
the RV platform, and otherwise just declares some external symbol.
This makes the code very difficult to test and mock because the
OT_PLATFORM_RV32 define is only true when compiling for RISC-V
but we do want to test some of this code on the host in a unittest.

Since the initial motivation for this was probably just to make it
inline, and we have since then enabled LTO, there is no realy reason
to do this anymore. This commit moves the device implementation to
ibex.c so that the header always just defines the prototypes.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>