[cryptotest] Allow longer RSA signatures
Some tests in the wycheproof test vector set are error tests. In one
of these error tests, a larger signature (514 instead of 512-bytes) is
sent to the DUT. However, we currently limit the signature payload to 512-
bytes, letting the test fail.

The solution to this is simple, increase the max. number of signature bytes
to 512-bytes. The CL now gets this 514-byte signature and internally fails,
as expected by the test.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>