[silicon_creator,pwrmgr] Create method to enable watchdog reset
Currently the library exports a very generic method to search
for a bitmask in the DT and it requires users to manually poke at
the pwrmgr registers. This is not really in the spirit of the
silicon_creator libraries and it can be misused.

This commit makes the search function internal and create a specific
method to enable the watch reset request, which is the only user
currently.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>