[cryptotest] Initialize msg len for RSA
For error test cases, `otcrypto_rsa_decrypt` is expected to abort.

In this case, it could be that `msg_len` is not set.  However, as
we use this variable for the memcpy below, we should initialize it
to 0 to avoid that we are corrupting memory.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>
(cherry picked from commit 702ae87891fe05fa4d3c9ee0de8638e28c62101f)