[cryptotest] Fix bug in RSA testvector parser
We only should abort parsing once we have iterated over all test_groups. Although
this is a bug in the `wycheproof_rsa_parser` there was no impact on RSA testing
as the wycheproof testvectors for RSA only consist of a single group.

However, it is better to fix this as it could be that future test vectors could
have multiple groups.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>