[csrng] Update IP description, regfile, and doc
This commit removes all now unused error reporting bits from the regfile
and documentation as well as the dummy-assigns in csrng_core.

It also changes the misleadingly named `CMD_GEN_CNT_ERR` bit to
`CTR_ERR` which reflects its meaning (collective reporting if _any_
hardened counter in CSRNG has reported an error).

Finally, it aligns the SEC_CM annotations in RTL to match the IP hjson
and moves some of the labels closer to where the actual counter measure
is implemented.

Signed-off-by: Florian Glaser <glaserf@lowrisc.org>