[racl_ctrl,dv] Move the RAL type from a param to ral_type_name
Now that the scoreboard uses a weaker RAL type in the parameter, we
can now switch the env_cfg to use it too.

This uses the framework developed a few commits ago ("Allow
dv_base_env_cfg with a templated RAL type"). The point is that we'll
now be able to template racl_ctrl_env_cfg (so that it can define
ral_type_name) without having to template its base class.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>