[racl_ctrl,dv] Expunge the templated RAL name from racl_ctrl DV
The clever bit of the trick is to define a templated env cfg class,
which is a tiny extension of the real one, whose only modification is
that it sets ral_type_name to be a templated RAL type.

Annoyingly, we didn't actually have a package in which to define the
class, so this commit has to do a bit of extra legwork, defining a
templated package and renaming the previous version of the package to
have a "_base_" in the name.

But that's it! The "interesting" part of the commit is the bit where
we no longer set ral_type_name in the module that is now called
racl_ctrl_base_env_cfg, but do set it in the templated env
cfg (racl_ctrl_env_cfg).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>