[dv] Allow dv_base_env_cfg with a templated RAL type
This is motivated by blocks that are templated, so the RAL model is
itself generated by templated code. It would be nice to avoid having
to template everything that contains a RAL.

After this commit, you can make a (simple) templated subclass that
derives from dv_base_env_cfg. This was a little tricky, because the
type parameters have to match! The solution here is to allow a user
to use a subclass of RAL_T instead.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>