[dv] Tidy up _DV_MUBI_DIST
Without this change, Xcelium warns about the branch that "isn't taken"
if you randomise something with DV_MUBI4_DIST (and presumably the
other similar macros).

Taking DV_MUBI4_DIST for concreteness, we have MuBi4True = 5 and
MuBi4False = 9. It seems that Xcelium works by compiling both sides of
the branch before it decides which one to use. As a result, it sees a
weight being assigned to the range [10:4]. Unfortunately,
SystemVerilog expects in constraints to be written [low:high], so this
is a weight assigned to the empty set of values 10 <= x <= 4.

In our default setup, this causes an warning: not a bad idea if the
author has done something silly like defining their endpoints
backwards.

Fortunately, we can just implement min() and max() with ternaries and
squash the two cases together.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>